+++ ServerSide/database/migrations/2016_12_09_105652_create_customer_clients_table.php
@@ -16,7 +16,7 @@ class CreateCustomerClientsTable extends Migration {
 			$table->increments('id');
 
 			$table->string('secret_key')->unique();
-			$table->unsignedInteger('credit')->default(0);
+			$table->double('credit')->default(0);
 
 			$table->string('customer_name');
 			$table->string('phone_number')->unique();
